# apiv1swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "OpenWhisk REST API",
        "description": "API for OpenWhisk",
        "version": "0.1.0"
    },
    "produces": [
        "application/json"
    ],
    "basePath": "/api/v1",
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic"
        }
    },
    "security": [
        {
            "basicAuth": []
        }
    ],
...
    "paths": {
...
        "/namespaces/{namespace}/actions/{actionName}": {
            "parameters": [
                {
                    "name": "namespace",
                    "in": "path",
                    "description": "The entity namespace",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "actionName",
                    "in": "path",
                    "description": "Name of action to fetch",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Get action information",
                "description": "Get action information.",
                "operationId": "getActionByName",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Include action code in the result",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Returned action",
                        "schema": {
                            "$ref": "#/definitions/Action"
                        }
                    },
                    "401": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "403": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "404": {
                        "$ref": "#/responses/ItemNotFound"
                    },
                    "500": {
                        "$ref": "#/responses/ServerError"
                    }
                }
            },
            "put": {
                "tags": [
                    "Actions"
                ],
                "description": "Create or update an action",
                "summary": "Create or update an action",
                "operationId": "updateAction",
                "parameters": [
                    {
                        "name": "overwrite",
                        "in": "query",
                        "description": "Overwrite item if it exists. Default is false.",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "true",
                            "false"
                        ]
                    },
                    {
                        "name": "action",
                        "in": "body",
                        "description": "The action being updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ActionPut"
                        }
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Updated Action",
                        "schema": {
                            "$ref": "#/definitions/Action"
                        }
                    },
                    "400": {
                        "$ref": "#/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "403": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "409": {
                        "$ref": "#/responses/Conflict"
                    },
                    "413": {
                        "$ref": "#/responses/RequestEntityTooLarge"
                    },
                    "500": {
                        "$ref": "#/responses/ServerError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Actions"
                ],
                "description": "Delete an action",
                "summary": "Delete an action",
                "operationId": "deleteAction",
                "responses": {
                    "200": {
                        "$ref": "#/responses/DeletedItem"
                    },
                    "400": {
                        "$ref": "#/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "403": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "404": {
                        "$ref": "#/responses/ItemNotFound"
                    },
                    "409": {
                        "$ref": "#/responses/Conflict"
                    },
                    "500": {
                        "$ref": "#/responses/ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Actions"
                ],
                "description": "Invoke an action",
                "summary": "Invoke an action",
                "operationId": "invokeAction",
                "parameters": [
                    {
                        "name": "blocking",
                        "in": "query",
                        "description": "Blocking or non-blocking invocation. Default is non-blocking.",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "true",
                            "false"
                        ]
                    },
                    {
                        "name": "result",
                        "in": "query",
                        "description": "Return only the result of a blocking activation. Default is false.",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "true",
                            "false"
                        ]
                    },
                    {
                        "name": "timeout",
                        "in": "query",
                        "description": "Wait no more than specified duration in milliseconds for a blocking response. Default value and max allowed timeout are 60000.",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "The parameters for the action being invoked",
                        "required": false,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful activation"
                    },
                    "202": {
                        "$ref": "#/responses/AcceptedActivation"
                    },
                    "401": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "403": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "404": {
                        "$ref": "#/responses/ItemNotFound"
                    },
                    "408": {
                        "$ref": "#/responses/Timeout"
                    },
                    "500": {
                        "$ref": "#/responses/ServerError"
                    },
                    "502": {
                        "description": "Activation produced an application error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "KeyValue": {
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "description": "Any JSON value"
                }
            }
        },
...
        "ErrorMessage": {
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                }
            }
        },
        "ActionLimits": {
            "description": "Limits on a specific action",
            "properties": {
                "timeout": {
                    "type": "integer",
                    "format": "int32",
                    "description": "timeout in milliseconds",
                    "default": 60000
                },
                "memory": {
                    "type": "integer",
                    "format": "int32",
                    "description": "memory in megabytes",
                    "default": 256
                },
                "logs": {
                    "type": "integer",
                    "format": "int32",
                    "description": "log size in megabytes",
                    "default": 10
                }
            }
        },
...
        "Action": {
            "required": [
                "namespace",
                "name",
                "version",
                "publish",
                "exec",
                "limits"
            ],
            "properties": {
                "namespace": {
                    "type": "string",
                    "description": "Namespace of the item",
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "description": "Name of the item",
                    "minLength": 1
                },
                "version": {
                    "type": "string",
                    "description": "Semantic version of the item",
                    "minLength": 1
                },
                "publish": {
                    "type": "boolean",
                    "description": "Whether to publish the item or not"
                },
                "exec": {
                    "$ref": "#/definitions/ActionExec"
                },
                "annotations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValue"
                    },
                    "description": "annotations on the item"
                },
                "parameters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValue"
                    },
                    "description": "parameter bindings included in the context passed to the action"
                },
                "limits": {
                    "$ref": "#/definitions/ActionLimits"
                },
                "updated": {
                    "type": "integer",
                    "description": "Time when the action was updated"
                }
            }
        },
        "ActionPut": {
            "description": "A restricted Action view used when updating an Action",
            "properties": {
                "namespace": {
                    "type": "string",
                    "description": "Namespace of the item",
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "description": "Name of the item",
                    "minLength": 1
                },
                "version": {
                    "type": "string",
                    "description": "Semantic version of the item",
                    "minLength": 1
                },
                "publish": {
                    "type": "boolean",
                    "description": "Whether to publish the item or not"
                },
                "exec": {
                    "$ref": "#/definitions/ActionExec"
                },
                "annotations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValue"
                    },
                    "description": "annotations on the item"
                },
                "parameters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValue"
                    },
                    "description": "parameter bindings included in the context passed to the action"
                },
                "limits": {
                    "$ref": "#/definitions/ActionLimits"
                }
            }
        },
        "ActionExec": {
            "properties": {
                "kind": {
                    "type": "string",
                    "enum": [
                        "nodejs:6",
                        "nodejs:8",
                        "nodejs:default",
                        "python:2",
                        "python:3",
                        "python:default",
                        "php:7.1",
                        "php:7.2",
                        "swift:3.1.1",
                        "swift:4.1",
                        "java",
                        "java:default",
                        "blackbox",
                        "sequence"
                    ],
                    "description": "the type of action"
                },
                "code": {
                    "type": "string",
                    "description": "The code to execute when kind is not 'blackbox'"
                },
                "image": {
                    "type": "string",
                    "description": "container image name when kind is 'blackbox'"
                },
                "main": {
                    "type": "string",
                    "description": "main entrypoint of the action code"
                },
                "init": {
                    "type": "string",
                    "description": "optional zipfile reference when code kind is 'nodejs'"
                },
                "binary": {
                    "type": "boolean",
                    "description": "Whether the action has a binary attachment or not"
                },
                "components": {
                    "type": "array",
                    "description": "For sequence actions, the individual action components",
                    "items": {
                        "type": "string"
                    }
                }
            },
            "description": "definition of the action, such as javascript code or the name of a container"
        },
...
        "ActivationId": {
            "required": [
                "activationId"
            ],
            "properties": {
                "activationId": {
                    "type": "string"
                }
            }
        },
...
    },
    "responses": {
        "BadRequest": {
            "description": "Bad request",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
        "UnauthorizedRequest": {
            "description": "Unauthorized request",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
        "ServerError": {
            "description": "Server error",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
...
        "DeletedItem": {
            "description": "Deleted Item"
        },
        "ItemNotFound": {
            "description": "Item not found",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
        "Timeout": {
            "description": "Request timed out"
        },
        "Conflict": {
            "description": "Conflicting item already exists",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
        "AcceptedActivation": {
            "description": "Accepted activation request",
            "schema": {
                "$ref": "#/definitions/ActivationId"
            }
        },
...
        "RequestEntityTooLarge": {
            "description": "Request entity too large",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
...
    }
}